API

ZeroNuFit.Analysis.retrieve_real_fit_results

ZeroNuFit.Analysis.run_analysis

ZeroNuFit.Analysis.save_outputs

ZeroNuFit.Likelihood.build_likelihood_looping_partitions

ZeroNuFit.Likelihood.build_likelihood_per_partition

ZeroNuFit.Likelihood.build_likelihood_zero_obs_evts

ZeroNuFit.Likelihood.build_prior

ZeroNuFit.Likelihood.exp_stable

ZeroNuFit.Likelihood.gaussian_plus_lowEtail

ZeroNuFit.Likelihood.generate_data

ZeroNuFit.Likelihood.get_bkg_pdf

ZeroNuFit.Likelihood.get_mu_b

ZeroNuFit.Likelihood.get_mu_s

ZeroNuFit.Likelihood.get_mu_s_b

ZeroNuFit.Likelihood.get_signal_bkg_priors

ZeroNuFit.Likelihood.get_signal_pdf

ZeroNuFit.Likelihood.get_signal_prior_info

ZeroNuFit.Likelihood.get_stat_blocks

ZeroNuFit.Likelihood.norm_exponential

ZeroNuFit.Likelihood.norm_linear

ZeroNuFit.Likelihood.norm_uniform

ZeroNuFit.Likelihood.run_fit_over_partitions

ZeroNuFit.Plotting.fit_model

ZeroNuFit.Plotting.plot_correlation_matrix

ZeroNuFit.Plotting.plot_data

ZeroNuFit.Plotting.plot_fit_and_data

ZeroNuFit.Plotting.plot_marginal_distr

ZeroNuFit.Plotting.plot_two_dim_posteriors

ZeroNuFit.Utils.check_key

ZeroNuFit.Utils.event_is_contained

ZeroNuFit.Utils.generate_disjoint_uniform_samples

ZeroNuFit.Utils.get_bkg_info

ZeroNuFit.Utils.get_corr_info

ZeroNuFit.Utils.get_deltaE

ZeroNuFit.Utils.get_efficiency

ZeroNuFit.Utils.get_energy_scale_pars

ZeroNuFit.Utils.get_events

ZeroNuFit.Utils.get_global_mode

ZeroNuFit.Utils.get_marginalized_mode

ZeroNuFit.Utils.get_par_posterior

ZeroNuFit.Utils.get_partition_event_index

ZeroNuFit.Utils.get_partitions

ZeroNuFit.Utils.get_partitions_events

ZeroNuFit.Utils.get_partitions_new

ZeroNuFit.Utils.get_range

ZeroNuFit.Utils.get_settings

ZeroNuFit.Utils.inverse_uniform_cdf

ZeroNuFit.Utils.save_generated_samples

ZeroNuFit.Utils.save_results_into_json

analysis.jl

ZeroNuFit.Analysis.retrieve_real_fit_results — Function

retrieve_real_fit_results(config::Dict{String,Any})

Function which handeles generating of fake data and run a signal+background model fit over it.

Arguments

config::Dict{String,Any}: the fit configuration.

ZeroNuFit.Analysis.run_analysis — Function

run_analysis(config::Dict{String,Any}; output_path::String, toy_idx = nothing)

Function which handeles running analysis.

Arguments

config::Dict{String,Any}: the fit configuration.

output_path::String: the path to the output files folder.

toy_idx: identification index of the generated toy.

ZeroNuFit.Analysis.save_outputs — Function

save_outputs(partitions, events, part_event_index, samples, posterior, nuisance_info, config, output_path, fit_ranges;priors=nothing,par_names=nothing,toy_idx=nothing)

Function to plot and save results, as well as inputs.

Arguments

partitions: table of partitions.

events: list of events (=energies) in each partition.

part_event_index: index mapping events to the partitions.

samples: set of generated MCMC samples.

posterior: posterior distribution evaluated via PosteriorMeasure(likelihood, prior).

nuisance_info: dictionary with info on the prior parameters.

config: input dictionary.

output_path: output folder path.

fit_ranges: dictionary of energy ranges considered for the analysis.

priors: prior distributions.

par_names: collection of parameter names.

toy_idx: identification index of the generated toy.

likelihood.jl

ZeroNuFit.Likelihood.norm_uniform — Function

norm_uniform(x::Real,p::NamedTuple,fit_range)

Normalised flat function defined by 1/norm.

Arguments

x::Real: the x value to evaluate at.

p::NamedTuple: collection of key-value pairs where each key corresponds to a model parameter.

fit_range: array of arrays, defining the allowed energy ranges; e.g. fit_range= [[1930,1950], [1970,1990], [2000,2050]].

ZeroNuFit.Likelihood.norm_linear — Function

norm_linear(x::Float64,p::NamedTuple,b_name::Symbol,fit_range)

Normalised linear function defined by (1+slope*(x-center)/net_width)/norm.

Arguments

x::Real: the x value to evaluate at.

p::NamedTuple: collection of key-value pairs where each key corresponds to a model parameter.

b_name::Symbol: name of the background index.

fit_range: array of arrays, defining the allowed energy ranges; e.g. fit_range= [[1930,1950], [1970,1990], [2000,2050]].

ZeroNuFit.Likelihood.norm_exponential — Function

norm_exponential(x::Float64,p::NamedTuple,b_name::Symbol,fit_range)

Normalised exponential function defined by exp_stable((x-center)*Rt)/norm.

Arguments

x::Real: the x value to evaluate at.

p::NamedTuple: collection of key-value pairs where each key corresponds to a model parameter.

b_name::Symbol: name of the background index.

fit_range: array of arrays, defining the allowed energy ranges; e.g. fit_range= [[1930,1950], [1970,1990], [2000,2050]].

ZeroNuFit.Likelihood.exp_stable — Function

exp_stable(x::Float64)

Exponential function, using Taylor expansion series for abs(x) < 1E-6.

Arguments

x::Real: the x value to evaluate at.

ZeroNuFit.Likelihood.gaussian_plus_lowEtail — Function

gaussian_plus_lowEtail(evt_energy::Float64,Qbb::Float64,bias::Float64,reso::Float64,part_k::NamedTuple)

Signal model based on the peak shape used for the MJD analysis. The peak shape was derived from considerations made in [S. I. Alvis et al., Phys. Rev. C 100, 025501 (2019)].

Arguments

evt_energy::Float64: energy of the event at which we want to compute the function.

Qbb::Float64: centroid of the Gaussian.

bias::Float64: energy bias associated to the energy event.

reso::Float64: energy resolution associated to the energy event.

part_k::NamedTuple: Table of specifications for a given partition k.

ZeroNuFit.Likelihood.get_mu_b — Function

get_mu_b(deltaE, exposure, bkg_index)

Get the expected number of background counts.

Arguments

deltaE: net width of the fit range.

exposure: exposure (mass x time).

bkg_index: background index value.

ZeroNuFit.Likelihood.get_mu_s — Function

get_mu_s(exposure, eff, signal)

Get the expected number of signal counts.

Arguments

exposure: exposure (mass x time).

eff: total signal efficiency.

signal: number of signal counts

ZeroNuFit.Likelihood.get_mu_s_b — Function

get_mu_s_b(p::NamedTuple,part_k::NamedTuple,idx_part_with_events::Int,settings::Dict,fit_range)

Get the expected number of signal and background counts.

Arguments

p::NamedTuple: collection of key-value pairs where each key corresponds to a model parameter.

part_k::NamedTuple: Table of specifications for a given partition k.

idx_part_with_events::Int: index of the partition with the event.

settings::Dict: dictionary of settings containing configuration for the likelihood calculation.

fit_range: array of arrays, defining the allowed energy ranges; e.g. fit_range= [[1930,1950], [1970,1990], [2000,2050]].

ZeroNuFit.Likelihood.build_likelihood_zero_obs_evts — Function

build_likelihood_zero_obs_evts(part_k::NamedTuple, p::NamedTuple,settings::Dict,fit_range)

Function to calculate the likelihood for a single data partition k with 0 events.

Arguments

part_k::NamedTuple: Table of specifications for a given partition k.

p::NamedTuple: collection of key-value pairs where each key corresponds to a model parameter.

settings::Dict: dictionary of settings containing configuration for the likelihood calculation.

fit_range: array of arrays, defining the allowed energy ranges; e.g. fit_range= [[1930,1950], [1970,1990], [2000,2050]].

ZeroNuFit.Likelihood.build_likelihood_per_partition — Function

build_likelihood_per_partition(idx_part_with_events::Int,part_k::NamedTuple, events_k::Vector{Union{Float64}},p::NamedTuple,settings::Dict,bkg_shape::Symbol,fit_range)

Function which computes the likelihood for a single data partition k.

Arguments

idx_part_with_events::Int: index of the partition with the event.

part_k::NamedTuple: Table of specifications for a given partition k.

events_k::Vector{Union{Float64}}: vecotr of events (=energies) in the partition k.

p::NamedTuple: collection of key-value pairs where each key corresponds to a model parameter.

settings::Dict: dictionary of settings containing configuration for the likelihood calculation.

bkg_shape::Symbol: Specifies the background shape; default is :uniform.

fit_range: array of arrays, defining the allowed energy ranges; e.g. fit_range= [[1930,1950], [1970,1990], [2000,2050]].

ZeroNuFit.Likelihood.build_likelihood_looping_partitions — Function

build_likelihood_looping_partitions(partitions::TypedTables.Table,events::Array{Vector{Float64}},part_event_index::Vector{Int},settings::Dict,sqrt_prior::Bool,s_max::Union{Float64,Nothing},fit_ranges;bkg_shape::Symbol=:uniform)

Function to build the likelihood (a DensityInterface.logfuncdensity object) for the fit by looping over partitions.

Arguments

partitions::TypedTables.Table: table of partitions.

events::Array{Vector{Float64}}: list of events (=energies) in each partition.

part_event_index::Vector{Int}: index mapping events to the partitions.

settings::Dict: A dictionary of settings containing configuration for the likelihood calculation.

sqrt_prior::Bool: Whether to include the square root prior in the likelihood calculation. If False, a uniform prior is used.

s_max::Union{Float64, Nothing}: A maximum value used for scaling the square root prior. If Nothing, no prior is applied.

fit_ranges: The fitting ranges corresponding to the partitions.

bkg_shape::Symbol: Specifies the background shape; default is :uniform.

ZeroNuFit.Likelihood.get_stat_blocks — Function

get_stat_blocks(partitions,events::Array{Vector{Float64}},part_event_index,fit_ranges;config,bkg_only::Bool)

Function to retrieve useful pieces (prior, likelihood, posterior), also in saving values.

Arguments

partitions: table of partitions.

events::Array{Vector{Float64}}: list of events (=energies) in each partition.

part_event_index::Vector{Int}: index mapping events to the partitions.

fit_ranges: dictionary of energy ranges considered for the analysis.

config: input dictionary.

bkg_only::Bool: True if we are using a model with background only.

ZeroNuFit.Likelihood.run_fit_over_partitions — Function

run_fit_over_partitions(partitions,events::Array{Vector{Float64}},part_event_index::Vector{Int}, config,fit_ranges)

Function to run the fit looping over partitions.

Arguments

partitions: table of partitions.

events::Array{Vector{Float64}}: list of events (=energies) in each partition.

part_event_index::Vector{Int}: index mapping events to the partitions.

config: input dictionary.

fit_ranges: dictionary of energy ranges considered for the analysis.

ZeroNuFit.Likelihood.get_signal_prior_info — Function

get_signal_prior_info(bkg_only::Bool,config)

Function that retrieves signal prior information.

Arguments

bkg_only::Bool: True if we are using a model with background only.

config: input dictionary.

ZeroNuFit.Likelihood.get_signal_bkg_priors — Function

get_signal_bkg_priors(config)

Defines specific priors for signal and background contributions.

Arguments

config: input dictionary.

ZeroNuFit.Likelihood.get_bkg_pdf — Function

get_bkg_pdf(bkg_shape::Symbol,evt_energy::Float64,p::NamedTuple,b_name::Symbol,fit_range)

Returns the background modeling function.

Arguments

bkg_shape::Symbol: Specifies the background shape; default is :uniform.

evt_energy::Float64: energy of the event at which we want to compute the function.

p::NamedTuple: collection of key-value pairs where each key corresponds to a model parameter.

b_name::Symbol: name of the background index.

fit_range: array of arrays, defining the allowed energy ranges; e.g. fit_range= [[1930,1950], [1970,1990], [2000,2050]].

ZeroNuFit.Likelihood.get_signal_pdf — Function

get_signal_pdf(evt_energy::Float64, Qbb::Float64, part_k::NamedTuple)

Returns the signal modeling function.

Arguments

evt_energy::Float64: energy of the event at which we want to compute the function.

Qbb::Float64: centroid of the Gaussian.

part_k::NamedTuple: Table of specifications for a given partition k.

ZeroNuFit.Likelihood.build_prior — Function

build_prior(partitions,part_event_index,config,settings::Dict;hierachical=false,hierachical_mode=nothing,hierachical_range=nothing,bkg_shape=:uniform,shape_pars=nothing)

Builds the priors for use in the fit.

Arguments

partitions::Table: Table of the partition info.

part_event_index::Vector{Int}: index for the parameters for partitions with events.

config: input dictionary.

settings::Dict: A dictionary of settings containing configuration for the likelihood calculation.

hierachical: True if we use a hierarchical model.

hierachical_mode: average mode of the hierarchical model.

hierachical_range: range of the reference distribution used to correlated the background indexes in a hierarchical model.

bkg_shape::Symbol: Specifies the background shape; default is :uniform.

shape_pars: background shape parameters, e.g. slope for linear or exponential background modeling.

ZeroNuFit.Likelihood.generate_data — Function

generate_data(samples::BAT.DensitySampleVector,partitions::TypedTables.Table,part_event_index::Vector{Int},settings::Dict,fit_ranges;best_fit::Bool=false,seed=nothing,bkg_only=false)

Generates data from a posterior distribution. This is based on the posterior predictive distributions. Given a model with some parameters theta_i, the posterior predictive distribution, or the distribution of data generated according to the posterior distribution of theta and the likelihood is:

\[p(y|D) =int p(y|theta)p(theta|D)dtheta\]

Or in terms of sampling we first draw samples of theta from the posterior and then generate, datasets based on the likelihood. We also give the options to fix the posterior distribution to the best fit, which is equivalent to the standard sampling methods.

Arguments

samples::DensitySamplesVector: samples of a past fit or a NamedTuple of best fit.

partitions::Table: Table of the partition info.

part_event_index::Vector{Int}: index for the parameters for partitions with events.

settings::Dict: A dictionary of settings containing configuration for the likelihood calculation.

fit_ranges: dictionary of energy ranges considered for the analysis.

best_fit::Bool: True if you want to fix the paramaters to the best fit.

seed::Int: random seed.

bkg_only::Bool: True if we are using a model with background only.

plotting.jl

ZeroNuFit.Plotting.plot_data — Function

plot_data(hist::Histogram,name,partitions,part_event_index,pars,samples,posterior,plotflag,settings::Dict,bkg_shape::Symbol,fit_ranges)

Function to plot events in the Qbb analysis window and BAT fit results.

Arguments

hist::Histogram: binned histogram with energy events.

name: title of the plot.

partitions: table of partitions.

part_event_index: index mapping events to the partitions.

pars: free parameters in format (:B, :S, ...)

samples: set of generated MCMC samples.

posterior: posterior distribution evaluated via PosteriorMeasure(likelihood, prior).

plotflag: True if the 68% band is plotted together with the best fit result.

settings::Dict: dictionary of settings containing configuration for the likelihood calculation.

bkg_shape::Symbol: Specifies the background shape; default is :uniform.

fit_ranges: The fitting ranges corresponding to the partitions.

ZeroNuFit.Plotting.plot_fit_and_data — Function

plot_fit_and_data(partitions,part_event_index,samples,posterior,pars,output,config,fit_ranges;toy_idx=nothing)

Retrieves event energies and calls plot_data to plot events in the Qbb analysis window and BAT fit results.

Arguments

partitions: table of partitions.

part_event_index: index mapping events to the partitions.

samples: set of generated MCMC samples.

posterior: posterior distribution evaluated via PosteriorMeasure(likelihood, prior).

pars: free parameters in format (:B, :S, ...)

output: output folder path.

config: input dictionary.

fit_ranges: The fitting ranges corresponding to the partitions.

toy_idx: identification index of the generated toy.

ZeroNuFit.Plotting.plot_correlation_matrix — Function

plot_correlation_matrix(samples,output;par_names=nothing,toy_idx=nothing)

Plots the correlation matrixes across the free parameters.

Arguments

samples: set of generated MCMC samples.

output: output folder path.

par_names: collection of parameter names.

toy_idx: identification index of the generated toy.

ZeroNuFit.Plotting.plot_marginal_distr — Function

plot_marginal_distr(partitions,samples,pars,output;sqrt_prior=false,priors=nothing,par_names=nothing,plot_config=nothing,s_max=nothing,hier=false,toy_idx=nothing)

Function to plot 1D and 2D marginalized posterior distributions, and prior distributions.

Arguments

partitions: table of partitions.

samples: set of generated MCMC samples.

pars: free parameters in format (:B, :S, ...)

output: output folder path.

sqrt_prior: True if a 1/sqrt(signal) prior distribution was used for the signal.

priors: True if we want to overlap priors distributions over posterior distributions.

par_names: collection of parameter names.

plot_config: dictionary with plotting settings.

s_max: maximum value of the marginalized signal posterior pdf.

hier: True if we use a hierarchical model.

toy_idx: identification index of the generated toy.

ZeroNuFit.Plotting.plot_two_dim_posteriors — Function

plot_two_dim_posteriors(samples,pars,output;par_names=nothing,toy_idx=nothing)

Plots the correlation matrixes across the free parameters.

Arguments

samples: set of generated MCMC samples.

pars: free parameters in format (:B, :S, ...) free parameters in format (:B, :S, ...)

output: output folder path.

par_names: collection of parameter names.

toy_idx: identification index of the generated toy.

ZeroNuFit.Plotting.fit_model — Function

fit_model(idx_part_with_events::Int,part_k::NamedTuple,p::NamedTuple,settings::Dict,bkg_shape::Symbol,fit_range,x::Float64)

Evaòuates the total modeling function at x.

Arguments

idx_part_with_events::Int: index of the partition with the event.

part_k::NamedTuple: Table of specifications for a given partition k.

p::NamedTuple: collection of key-value pairs where each key corresponds to a model parameter.

settings::Dict: dictionary of settings containing configuration for the likelihood calculation.

bkg_shape::Symbol: Specifies the background shape; default is :uniform.

fit_range: array of arrays, defining the allowed energy ranges; e.g. fit_range= [[1930,1950], [1970,1990], [2000,2050]].

x::Float64: the x value to evaluate at.

utils.jl

ZeroNuFit.Utils.check_key — Function

check_key(config::Dict, k::String)

Function that checks the existence of a key in a dictionary. If the key is not found, the code exits here.

Arguments

config::Dict: input dictionary.

k::String: name of the key to check the existence of in config.

ZeroNuFit.Utils.get_settings — Function

get_settings(config::Dict{String,Any})

Function that retrieves useful settings information from the input configuration dictionary.

Returns a dictionary containing information on energy bias/resolution/efficiency (if fixed or not, if correlated or not) and on the type of fit (if background only or not).

Arguments

config::Dict{String,Any}: input dictionary.

ZeroNuFit.Utils.get_events — Function

get_events(event_path::String,partitions)::Array{Vector{Float64}}

Function that returns an Array of Vectors filled with energy events per each partition. The code exits here if an event can't be associated to any existing partition.

Arguments

event_path::String: path to the input JSON file with energy events.

partitions: Table of retrieve partitions.

ZeroNuFit.Utils.event_is_contained — Function

event_is_contained(event::Float64, fit_ranges)

Function to check the containment of an energy event.

Returns true if the event is contained at least in one of the selected energy ranges.

Arguments

event::Float64: energy event.

fitrange: array of arrays, defining the allowed energy ranges; e.g.fitrange= [[1930,1950], [1970,1990], [2000,2050]]`

ZeroNuFit.Utils.get_partitions — Function

get_partitions(config::Dict{String, Any})

Function to retrieve a Table of partitions and a dictionary of fit ranges.

Arguments

config::Dict{String,Any}: input dictionary.

ZeroNuFit.Utils.get_partitions_new — Function

get_partitions_new(part_path::String)

Get the partition information from a JSON file and save to a Table.

Returns a Table of partitions, a dictionary of fit groups, a dictionary of fit ranges.

Arguments

part_path::String: path to a given partition JSON file.

ZeroNuFit.Utils.get_partitions_events — Function

get_partitions_events(config::Dict{String, Any})

Get partition, event, and fit range info from the configuration dictionary given in input.

Returns an object descirbing if a partition has an event by assigning indexes, an array of energy events, a Table of partitions, and fit ranges.

Arguments

config::Dict{String,Any}: input dictionary.

ZeroNuFit.Utils.get_partition_event_index — Function

get_partition_event_index(events::Array{Vector{Float64}},partitions::TypedTables.Table)::Vector{Int}

Returns an object describing if a partition has an event and giving them indexes. This creates a vector where

V[i]=0 if partition i has no events,

V[i]=idx if partition i has events,

where the index counts the number of partitions with index<=i.

Arguments

events::Array{Vector{Float64}}: list of events (=energies) in each partition.

partitions::TypedTables.Table: table of partitions.

ZeroNuFit.Utils.get_corr_info — Function

get_corr_info(config)

Function that retrieves information about correlated background from config in input.

Arguments

config: input dictionary.

ZeroNuFit.Utils.get_energy_scale_pars — Function

get_energy_scale_pars(part_k::NamedTuple,p::NamedTuple,settings::Dict,idx_part_with_events)

Returns the energy resolution and bias for a given partition, depending on the specified settings (e.g. if correlated or not, if fixed or not).

Arguments

p::NamedTuple: collection of key-value pairs where each key corresponds to a model parameter.

part_k::NamedTuple: Table of specifications for a given partition k.

settings::Dict: a dictionary containing information on energy bias/resolution/efficiency (if fixed or not, if correlated or not) and on the type of fit (if background only or not).

idx_part_with_events::Int: index of the partition with the event.

ZeroNuFit.Utils.get_efficiency — Function

get_efficiency(p::NamedTuple,part_k::NamedTuple,idx_part_with_events::Int,settings::Dict)

Returns the efficiency for a given partition, depending on the specified settings (e.g. if correlated or not, if fixed or not). If you are fitting with a background-only model, then efficiency=0.

Arguments

p::NamedTuple: collection of key-value pairs where each key corresponds to a model parameter.

part_k::NamedTuple: Table of specifications for a given partition k.

idx_part_with_events::Int: index of the partition with the event.

settings::Dict: a dictionary containing information on energy bias/resolution/efficiency (if fixed or not, if correlated or not) and on the type of fit (if background only or not).

ZeroNuFit.Utils.get_deltaE — Function

get_deltaE(fit_range)

Function that returns the net width of the fit range.

Arguments

fit_range: array of arrays, defining the allowed energy ranges; e.g. fit_range= [[1930,1950], [1970,1990], [2000,2050]].

ZeroNuFit.Utils.get_range — Function

get_range(fit_range)

Function that returns lower and upper edges of fit ranges.

Arguments

fit_range: array of arrays, defining the allowed energy ranges; e.g. fit_range= [[1930,1950], [1970,1990], [2000,2050]].

ZeroNuFit.Utils.get_bkg_info — Function

get_bkg_info(config)

Function that retrieves background shape name and parameters (if different from flat) from the input configuration dictionary.

Arguments

config: input dictionary.

ZeroNuFit.Utils.get_global_mode — Function

get_global_mode(samples, posterior)

Function which retrieves global mode and a refined estimate of it (using bat_findmode).

Arguments

samples: set of generated MCMC samples.

posterior: posterior distribution evaluated via PosteriorMeasure(likelihood, prior).

ZeroNuFit.Utils.get_marginalized_mode — Function

get_marginalized_mode(samples, par)

Function which retrieves marginalized mode as the highest bin of the posterior, using 250 bins (vs 100 bins set by default by BAT).

Arguments

samples: set of generated MCMC samples.

par: name of the parameter for which we want to extract the marginalized mode.

ZeroNuFit.Utils.get_par_posterior — Function

get_par_posterior(samples, par; idx)

Function that retrieves the parameter posterior.

Arguments

samples: set of generated MCMC samples.

par: name of the parameter for which we want to extract the marginalized mode.

idx: index for multiparameters

ZeroNuFit.Utils.inverse_uniform_cdf — Function

inverse_uniform_cdf(p, fit_range)

Returns the inverse cumulative distribution function value for the given probability p.

Arguments

p: probability between 0 and 1 representing the desired quantile of the cumulative distribution.

fit_range: array of arrays, defining the allowed energy ranges; e.g. fit_range= [[1930,1950], [1970,1990], [2000,2050]].

ZeroNuFit.Utils.generate_disjoint_uniform_samples — Function

generate_disjoint_uniform_samples(n, fit_range; seed = nothing)

Generates a list of n events uniform sampled within a specified range using the inverse CDF method.

Arguments

n: number of events to randomly generate.

fit_range: array of arrays, defining the allowed energy ranges; e.g. fit_range= [[1930,1950], [1970,1990], [2000,2050]].

seed: fix to a value if you want to fix the random generator seed.

ZeroNuFit.Utils.save_generated_samples — Function

save_generated_samples(samples,output)

Saves generated MCMC samples to .jld2 and .h5 files.

Arguments

samples: set of generated MCMC samples.

output: output folder path.

ZeroNuFit.Utils.save_results_into_json — Function

save_results_into_json(samples,posterior,nuisance_info,config,output;par_names=nothing,toy_idx=nothing)

Function which saves results from the fit and the used input configurations.

Arguments

samples: set of generated MCMC samples.

posterior: posterior distribution evaluated via PosteriorMeasure(likelihood, prior).

nuisance_info: dictionary with info on the prior parameters.

config: input dictionary.

output: output folder path.

par_names: collection of parameter names.

toy_idx: identification index of the generated toy.